Change authid to float | experimenting

+++ db/migrate/20140513145917_convert_authid_int_to_double_in_users.rb
@@ -1,6 +1,6 @@
 class ConvertAuthidIntToDoubleInUsers < ActiveRecord::Migration
   def up
-  	change_column :users, :authid, :double
+  	change_column :users, :authid, :float
   end
 
   def down
